+++ R/tidyrisk_scenario.R
@@ -1,5 +1,8 @@
 #' Construct a quantitative scenario object
 #'
+#' Supply one or more named lists in the format of `foo_params`,
+#' where each `foo` is an OpenFAIR element name (e.g. tef, tc, diff, lm).
+#'
 #' @param tef_params List of threat frequency parameters
 #' @param tc_params List of threat capability parameters
 #' @param diff_params List of difficulty parameters
@@ -14,12 +17,14 @@ new_tidyrisk_scenario <- function(tef_params = list(),
   stopifnot(is.list(tef_params), is.list(tc_params),
             is.list(diff_params), is.list(lm_params),
             is.character(model))
-  scenario <- list(tef_params = tef_params,
-                   tc_params = tc_params,
-                   diff_params = diff_params,
-                   lm_params = lm_params,
-                   model = model)
-  class(scenario) <- c("tidyrisk_scenario") #class(scenario))
+  scenario <- list(
+    parameters = list(
+      tef = tef_params,
+      tc = tc_params,
+      diff = diff_params,
+      lm = lm_params),
+    model = model)
+  class(scenario) <- c("tidyrisk_scenario")
   scenario
 }
 
@@ -31,11 +36,39 @@ tidyrisk_scenario <- function(tef_params = list(), tc_params = list(),
   new_tidyrisk_scenario(tef_params, tc_params, diff_params, lm_params, model)
 }
 
+#' Default printing of a tidyrisk_scenario
+#'
+#' Basic printing of a tidyrisk scenario
+#'
+#' @param x A tidyrisk_scenario
+#' @param ... Currently not used
 #' @export
 #' @importFrom cli cat_line
 print.tidyrisk_scenario <- function(x, ...) {
   cli::cat_line("# Scenario model: ", x$model)
-  NextMethod(x)
+  cli::cat_line("# Defined parameters: ", paste0(names(x$parameters), collapse = ", "))
+  #NextMethod(x)
 
   invisible(x)
+}
+
+#' Test if the object is a tidyrisk_scenario
+#'
+#' This function returns `TRUE` for tidyrisk_scenario (or subclasses)
+#' and `FALSE` for all other objects.
+#'
+#' @param x An object
+#' @return `TRUE` if the object inherits from the `tidyrisk_scenario` class.
+#' @export
+is_tidyrisk_scenario <- function(x) {
+  inherits(x, "tidyrisk_scenario")
+}
+
+#' Alias test for tidyrisk scenario
+#'
+#' Please use `is_tidyrisk_scenario()` instead.
+#' @param x An object
+#' @export
+is.tidyrisk_scenario <- function(x) {
+  is_tidyrisk_scenario(x)
 }